requirements.txt
flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.2.0
joblib>=1.2.0
numpy>=1.23.0
Pillow>=9.5.0
